backend/pyproject.toml
[project]
name = "market-signal-hub-backend"
version = "0.1.0"
description = "Economic/news/community collection and analytics API"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "apscheduler>=3.10.4",
  "asyncpg>=0.30.0",
  "beautifulsoup4>=4.12.3",
  "black>=25.1.0",
  "feedparser>=6.0.11",
  "fastapi>=0.115.0",
  "httpx>=0.28.1",
  "orjson>=3.10.16",
  "psycopg[binary]>=3.2.6",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.7.1",
  "python-dateutil>=2.9.0.post0",
  "python-slugify>=8.0.4",
  "ruff>=0.11.0",
  "sqlalchemy>=2.0.38",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["B008"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.uv]
package = false
